requirements.txt
pynmea2
pyserial
paho-mqtt